Full BiorXiv version, with updated README and licensing information.
Additional copyedits, layout changes, and small figure changes.
Additionally, created a long README that describes how to simulate
your own systems, in addition to information on how to replicate
all of the information in the figures.

Licensing changed to be clear that all text/figures are currently under
CC-BY-SA-NC (until publication), but the rest of the code is
MIT-licensed.